Move cross-version test to appropriate subproject
