+++ subprojects/language-java/src/crossVersionTest/groovy/org/gradle/api/internal/tasks/compile/tooling/JavaCompileTaskSuccessResultCrossVersionTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.tooling.r51
+package org.gradle.api.internal.tasks.compile.tooling
 
 import org.gradle.integtests.tooling.fixture.ProgressEvents
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
