+++ subprojects/tooling-api/tooling-api.gradle.kts
@@ -70,7 +70,6 @@ testFixtures {
     from(":logging")
     from(":dependencyManagement")
     from(":ide")
-    from(":languageJava")
 }
 
 apply(from = "buildship.gradle")
